WIP: Allow reusing existing kernel tree as-is
This skips all git checkout, patching, configuring and cleaning when
USE_EXISTING_KERNEL_TREE=yes is passed.

This can be useful when doing kernel development by making changes
directly to the kernel sourcetree in `cache/sources`, while still
allowing to use the `compile.sh` to use the same compilation commands,
deb package building, ccache, etc. It is also useful when bisecting
a kernel issue.

Can be used something like this:

	USE_EXISTING_KERNEL_TREE=yes ./compile.sh default kernel

To produce a kernel deb in a few minutes ready to be installed on
a device.

This is mostly a proof of concept to explore the usecase. In particular:
 - I am not happy with the USE_EXISTING_KERNEL_TREE name yet.
 - I wonder if this variable should maybe be split into different ones
   (e.g. for skipping git checkout, patching, config, clean already
   has one that can be used).
 - Or maybe it would be better to split off some code out of
   compile_kernel() and expose a kernel_build and/or kernel_package
   command from ./compile.sh instead?